refactor(desktop): share branch combobox, drop native select
Promote the search-as-you-type branch picker from NewSessionDialog into
a shared component under features/worktree and reuse it in
SessionSettingsDialog — the native select there clashed with the rest
of the redesigned UI. Adds an excludeNames prop and an openDirection
knob for callers near the bottom of a scroll container.
